убрала интерфейс для модели, сделала reparedStatement для существующих вставок

+++ src/main/java/com/skytecgames/testtask/sokolova/model/impl/Clan.java
@@ -7,7 +7,7 @@ import lombok.Getter;
 @AllArgsConstructor
 @Getter
 public class Clan {
-    private long id;     // id клана
+    private int id;     // id клана
     private String name; // имя клана
     private int gold;    // текущее количество золота в казне клана
